+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # TelegramBotAPI changelog
 
+## 3.2.6
+
+* `Core`:
+    * Fixes in `ChatMemberSerializer#serialize` method
+    * Migration of `EntitiesBuilder` from `Utils` to `Core`
+* `API`:
+    * All the extensions related to text messages (replies, sending, editing) got their duplicates with `buildEntities` lambda and separator
+* `Utils`:
+    * Migration of `EntitiesBuilder` from `Utils` to `Core`
+
 ## 3.2.5
 
 * `Common`: